abstracted to add config.ini

+++ faucet.py
@@ -72,14 +72,14 @@ def send_faucet_transaction(address: str, tokens: float):
 def send_mumbai_faucet_transaction(address: str, tokens: float):
     nonce = mumbai_w3.eth.getTransactionCount(FAUCET_ADDRESS)
 
-    for gas in [35 * 1e9, 50 * 1e9, 100 * 1e9, 350 * 1e9, 500 * 1e9, 1000 * 1e9]:
+    for gas in [int(35 * 1e9), int(50 * 1e9), int(100 * 1e9), int(350 * 1e9), int(500 * 1e9), int(1000 * 1e9)]:
         try:
             log("Trying testnet transaction to " + address + " with nonce " + str(nonce) + " and gas " + str(gas / 1e9))
 
             # Create the transaction
             signed_txn = mumbai_w3.eth.account.sign_transaction(dict(
                 nonce=nonce,
-                gasPrice=gas,
+                gasPrice=int(gas),
                 gas=50000,
                 to=address,
                 value=int(tokens * 1e18),
